refactor: user routes are kept in a single file
